requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
requests==2.31.0
celery
redis
python-dotenv
boto3
jinja2
python-multipart
itsdangerous
supabase
bcrypt